+++ system/Core/Controller.php
@@ -13,32 +13,14 @@ use Core\Language;
 use Core\Template;
 use Core\View;
 use Http\Response;
-
-use Symfony\Component\HttpFoundation\Response as SymfonyResponse;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-
-use Event;
+use Routing\Controller as BaseController;
 
 
 /**
  * Core controller, all other controllers extend this base controller.
  */
-abstract class Controller
+abstract class Controller extends BaseController
 {
-    /**
-     * The requested Method by Router.
-     *
-     * @var string|null
-     */
-    private $method = null;
-
-    /**
-     * The parameters given by Router.
-     *
-     * @var array
-     */
-    private $params = array();
-
     /**
      * The currently used Template.
      *
@@ -65,6 +47,8 @@ abstract class Controller
      */
     public function __construct()
     {
+        parent::__construct();
+
         // Setup the used Template to default, if it is not already defined.
         if(! isset($this->template)) {
             $this->template = Config::get('app.template');
@@ -72,37 +56,6 @@ abstract class Controller
 
         // Initialise the Language object.
         $this->language = Language::getInstance();
-    }
-
-    /**
-     * Execute the Controller Method
-     * @return \Symfony\Component\HttpFoundation\Response
-     *
-     * @throw \Exception
-     */
-    public function execute($method, $params = array())
-    {
-        // Initialise the Controller's variables.
-        $this->method = $method;
-        $this->params = $params;
-
-        // Notify the interested Listeners about the iminent Controller's execution.
-        Event::fire('framework.controller.executing', array($this, $method, $params));
-
-        // Before the Action execution stage.
-        $response = $this->before();
-
-        // When the Before Stage do not return a Symfony Response, execute the requested
-        // Method with the given arguments, capturing its returned value on our response.
-        if (! $response instanceof SymfonyResponse) {
-            $response = call_user_func_array(array($this, $method), $params);
-        }
-
-        // After the Action execution stage.
-        $this->after($response);
-
-        // Do the final post-processing stage and return the response.
-        return $this->processResponse($response);
     }
 
     /**
@@ -114,57 +67,33 @@ abstract class Controller
      */
     protected function processResponse($response)
     {
-        // If the response which is returned from the Controller's Action is null and we have
-        // View instances on View's Legacy support, we will assume that we are on Legacy Mode.
-        if (is_null($response)) {
-            // Retrieve the Legacy View instances.
-            $views = View::getLegacyViews();
-
-            if (! empty($views)) {
-                $headers = View::getLegacyHeaders();
-
-                // Fetch every View instance and append it to the Response content.
-                $content = '';
-
-                foreach ($views as $view) {
-                    $content .= $view->fetch();
-                }
-
-                // Create a Response instance from gathered information and return it.
-                return new Response($content, 200, $headers);
-            }
-        }
-
         // If the response which is returned from the Controller's Action is a View instance,
         // we will assume we want to render it using the Controller's templated environment.
-        if ($response instanceof View) {
-            if ($this->layout !== false) {
-                $response = Template::make($this->layout, $this->template)->with('content', $response);
-            }
+        if (($this->layout !== false) && ($response instanceof View)) {
+            return Template::make($this->layout, $this->template)->with('content', $response);
+        } else if (! is_null($response)) {
+            return $response;
         }
 
-        return $response;
-    }
+        // If the response which is returned from the Controller's Action is null and we have
+        // View instances on View's Legacy support, we will assume that we are on Legacy Mode.
 
-    /**
-     * Method automatically invoked before the current Action, stopping the flight
-     * when it returns false. This Method is supposed to be overriden for using it.
-     */
-    protected function before()
-    {
-        // Return true to continue the processing.
-        return true;
-    }
+        // Retrieve the Legacy View instances.
+        $views = View::getLegacyViews();
 
-    /**
-     * This method automatically invokes after the current Action and is supposed
-     * to be overriden for using it.
-     *
-     * Note that the Action's returned value is passed to this Method as parameter.
-     */
-    protected function after($result)
-    {
-        //
+        if (! empty($views)) {
+            $content = '';
+
+            $headers = View::getLegacyHeaders();
+
+            // Fetch every View instance and append it to the Response content.
+            foreach ($views as $view) {
+                $content .= $view->fetch();
+            }
+
+            // Create a Response instance from gathered information and return it.
+            return new Response($content, 200, $headers);
+        }
     }
 
     /**
@@ -229,49 +158,6 @@ abstract class Controller
     protected function getLayout()
     {
         return $this->layout;
-    }
-
-    /**
-     * @return mixed
-     */
-    protected function getMethod()
-    {
-        return $this->method;
-    }
-
-    /**
-     * @return array
-     */
-    protected function getParams()
-    {
-        return $this->params;
-    }
-
-    /**
-     * Handle calls to missing methods on the controller.
-     *
-     * @param  array   $parameters
-     * @return mixed
-     *
-     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
-     */
-    public function missingMethod($parameters = array())
-    {
-        throw new NotFoundHttpException("Controller method not found.");
-    }
-
-    /**
-     * Handle calls to missing methods on the controller.
-     *
-     * @param  string  $method
-     * @param  array   $parameters
-     * @return mixed
-     *
-     * @throws \BadMethodCallException
-     */
-    public function __call($method, $parameters)
-    {
-        throw new \BadMethodCallException("Method [$method] does not exist.");
     }
 
 }
